requirements.txt
# Core scientific stack
numpy
scipy
pillow
tqdm

# For testing
pytest

# For experiment configs/logging
pyyaml
